refactor(fuse): better table-meta and parquet reader function
