+++ src/query/storages/fuse/src/operations/mutation/deletion/deletion_source.rs
@@ -183,7 +183,9 @@ impl Processor for DeletionSource {
     fn process(&mut self) -> Result<()> {
         match std::mem::replace(&mut self.state, State::Finish) {
             State::FilterData(part, chunks) => {
-                let data_block = self.block_reader.deserialize(part.clone(), chunks)?;
+                let data_block = self
+                    .block_reader
+                    .deserialize_column_chunks_to_block(part.clone(), chunks)?;
                 let filter_result = self
                     .filter
                     .eval(&self.ctx.try_get_function_context()?, &data_block)?
@@ -223,7 +225,8 @@ impl Processor for DeletionSource {
                 let merged = if chunks.is_empty() {
                     data_block
                 } else if let Some(remain_reader) = self.remain_reader.as_ref() {
-                    let remain_block = remain_reader.deserialize(part, chunks)?;
+                    let remain_block =
+                        remain_reader.deserialize_column_chunks_to_block(part, chunks)?;
                     let remain_block = DataBlock::filter_block(remain_block, &filter)?;
                     for (col, field) in remain_block
                         .columns()
