+++ src/query/storages/fuse/src/operations/read/parquet_data_source_deserializer.rs
@@ -139,13 +139,15 @@ impl Processor for DeserializeDataTransform {
             let columns_chunks = read_res.columns_chunks()?;
             let part = FusePartInfo::from_part(&part)?;
 
-            let data_block = self.block_reader.deserialize_columns(
-                part.nums_rows,
-                &part.compression,
-                &part.columns_meta,
-                columns_chunks,
-                Some(self.uncompressed_buffer.clone()),
-            )?;
+            let data_block = self
+                .block_reader
+                .deserialize_column_chunks_to_block_with_buffer(
+                    part.nums_rows,
+                    &part.compression,
+                    &part.columns_meta,
+                    columns_chunks,
+                    Some(self.uncompressed_buffer.clone()),
+                )?;
 
             // Perf.
             {
