+++ src/query/storages/table-meta/src/meta/v1/snapshot.rs
@@ -21,7 +21,7 @@ use common_datavalues::DataSchema;
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::meta::common::FormatVersion;
+use crate::meta::statistics::FormatVersion;
 use crate::meta::ClusterKey;
 use crate::meta::Location;
 use crate::meta::SnapshotId;
